requirements.txt
hydra-core==1.3.2
matplotlib==3.10.3
numpy==2.2.6
omegaconf==2.3.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
streamlit==1.46.0
torch==2.7.0
torchinfo==1.8.0
torchvision==0.22.0
fastapi
uvicorn[standard]
torch
torchvision
pillow
mlflow
prometheus_client